requirements.txt
Flask==2.3.3
redis==5.0.1
rq==1.15.1
SQLAlchemy==2.0.23
PyMySQL==1.1.0
requests==2.31.0
python-slugify==8.0.1
boto3==1.34.0
cryptography==41.0.7